moving get_speedy_coords into speedy_coord.py
