+++ jcm/physics/speedy/humidity_test.py
@@ -17,9 +17,8 @@ class TestHumidityUnit(unittest.TestCase):
         from jcm.physics.speedy.humidity import get_qsat, spec_hum_to_rel_hum, rel_hum_to_spec_hum
         from jcm.forcing import ForcingData
         from jcm.physics.speedy.params import Parameters
-        from jcm.physics.speedy.utils import get_speedy_coords
         from jcm.terrain_data import TerrainData
-        from jcm.physics.speedy.speedy_coords import SpeedyCoords
+        from jcm.physics.speedy.speedy_coords import SpeedyCoords, get_speedy_coords
 
         parameters = Parameters.default()
         coords = get_speedy_coords(layers=kx, nodal_shape=(ix, il))
