+++ jcm/physics/speedy/surface_flux_test.py
@@ -15,9 +15,8 @@ class TestSurfaceFluxesUnit(unittest.TestCase):
         from jcm.physics.speedy.physics_data import SurfaceFluxData, HumidityData, ConvectionData, SWRadiationData, LWRadiationData, PhysicsData
         from jcm.physics_interface import PhysicsState, PhysicsTendency
         from jcm.physics.speedy.params import Parameters
-        from jcm.physics.speedy.utils import get_speedy_coords
         from jcm.terrain_data import TerrainData
-        from jcm.physics.speedy.speedy_coords import SpeedyCoords
+        from jcm.physics.speedy.speedy_coords import SpeedyCoords, get_speedy_coords
         from jcm.physics.speedy.test_utils import convert_to_speedy_latitudes
 
         parameters = Parameters.default()
@@ -466,9 +465,8 @@ class TestAquaplanetSurfaceFluxes(unittest.TestCase):
         from jcm.physics.speedy.params import Parameters
         from jcm.terrain_data import TerrainData
         from jcm.constants import grav
-        from jcm.physics.speedy.utils import get_speedy_coords
         from jcm.terrain_data import TerrainData
-        from jcm.physics.speedy.speedy_coords import SpeedyCoords
+        from jcm.physics.speedy.speedy_coords import SpeedyCoords, get_speedy_coords
         from jcm.physics.speedy.test_utils import convert_to_speedy_latitudes
 
         parameters = Parameters.default()
